+++ CHANGELOG.md
@@ -7,6 +7,14 @@ All notable changes to this project will be documented in this file, in reverse 
 - UpstreamAction reference
 - ADR on Postgres module with details on implementation specifics (MVCC, etc.)
 
+## 1.4.0 - 2020-09-25
+
+MariaDb integration
+
+## 1.3.0 - 2020-09-15
+
+Action retention policy
+
 ## 1.0.0 - 2020-05-19
 
 Initial release.